[add] cms board
