--- app/controllers/public_board/agents/tasks/node/posts_controller.rb
@@ -1,8 +1,0 @@
-class PublicBoard::Agents::Tasks::Node::PostsController < ApplicationController
-  include Cms::PublicFilter::Node
-
-  public
-    def generate
-      generate_node_with_pagination @node, max: 10
-    end
-end
