+++ app/models/board/file.rb
@@ -1,0 +1,11 @@
+class Board::File
+  include SS::Model::File
+  include SS::UserPermission
+  include SS::Relation::Thumb
+
+  default_scope ->{ where(model: "board/post") }
+
+  # TODO: add validation for public posted
+  private
+    #
+end
