--- app/models/public_board/file.rb
@@ -1,11 +1,0 @@
-class PublicBoard::File
-  include SS::Model::File
-  include SS::UserPermission
-  include SS::Relation::Thumb
-
-  default_scope ->{ where(model: "public_board/post") }
-
-  # TODO: add validation for public posted
-  private
-    #
-end
