+++ app/views/board/agents/addons/file/_show.html.erb
