+++ app/views/board/agents/addons/file_setting/_form.html.erb
