+++ app/views/board/agents/addons/list/_show.html.erb
