+++ app/views/board/agents/addons/post_setting/_show.html.erb
