+++ app/views/board/agents/nodes/post/_form.html.erb
