+++ app/views/board/agents/nodes/post/delete.html.erb
