+++ app/views/board/agents/nodes/post/index.html.erb
@@ -1,0 +1,20 @@
+<%= jquery do %>
+$(".toggle").on("click", function() {
+  $(this).closest(".post").find(".children").toggle();
+  return false;
+});
+$(".children").hide();
+<% end %>
+
+<%= @cur_node.upper_html.to_s.html_safe %>
+<nav class="menu">
+  <%= link_to t("board.views.new"), "new" %>
+  <%= link_to t("board.views.search"), "search" %>
+</nav>
+
+<% @items.each do |topic| %>
+  <%= render partial: "post", locals: { item: topic, children: (@cur_node.mode == "tree") }  %>
+<% end %>
+<%= @cur_node.lower_html.to_s.html_safe %>
+
+<%= paginate @items if @items.try(:current_page) %>
