+++ app/views/board/agents/nodes/post/new.html.erb
