+++ app/views/board/agents/nodes/post/search.html.erb
