+++ app/views/board/agents/nodes/post/sent.html.erb
