+++ app/views/board/main/_navi.html.erb
@@ -1,0 +1,5 @@
+<%= node_navi :board do %>
+
+<h3><%= link_to :"board.post", board_posts_path %></h3>
+
+<% end %>
