--- app/views/public_board/main/_navi.html.erb
@@ -1,10 +1,0 @@
-<%= node_navi do %>
-
-<nav class="sub-menu">
-  <header><h1><%= t "modules.public_board" %></h1></header>
-  <div>
-    <%= link_to :"public_board.post", public_board_posts_path, class: "icon-page" %>
-  </div>
-</nav>
-
-<% end %>
